smallint

+++ database/migrations/new/2021_02_01_073758_create_proposed_programs_table.php
@@ -16,7 +16,7 @@ class CreateProposedProgramsTable extends Migration
         Schema::create('proposed_programs', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedSmallInteger('user_id');
             $table->string('class')->nullable();
             $table->string('subject')->nullable();
             $table->string('modul')->nullable();
